BIG PROBLEM:
Okay, tried to implement line/column reporting scheme within the scanner
for compiler error reporting, among other things.  It did not work out, and
I found a vital design flaw in terms of how I've been handing thing on the
preprocessor end.

 * It turns out that including '#include' files into the main interim file,
   and them running further preprocessing on that file is not going to work.
   This is simply because by doing so, I modify the line numbers in file, and
   make it impossible to report on line numbers with certainty.

The alternative, by refaction of code, is to make it so that I can call
the "run_preprocessor" function on each file that is included, and never
insert lines of code into a file like that.  This will have to be a
recursive call of sorts.  Also, the run_scanner needs to work in the same
manner, on each included file.  So it seems that the inclusion or
"run_preprocessor" function, will not be able to handle comment removal.

I will need to consider this before I continue with coding, as it's vital
to later parts of error reporting, which among other things will help me
in debugging my compiler.