I can't believe how easy it was for me to modify the getword function to work with strings.  I have tested it in various cases, and it seems to work just fine.  It throws an error if the string is badly formated.
That is, if the string currentline does not end with the delimter
quotations marks, it exits.  The check looks for a newline before the
terminating quotes.

Of course, I see a much issues with this function in the future.
Particularly since it's so integral to everything that I am doing
with this scanner.  We will just have to wait and see.