Huge bug found.  I cannot figure out how to fix it.
It has to do with the handing of large strings.  For some
reason, the program crashes if the strings are
too large, and there is no space in a printf call, like:

printf("ver large strig");

I'll abandon this bug for now, maybe come back during the summer
for it.