I believe this is a buggy commit.  I will see if I have to fix bugs later, but for now it looks pretty good.  Here are some update:
Two major parser bugs discovered and fixed.

First, the declarations we messed up because of a simple scope_ptr
problem.  I had it set to -1;

Second, I fixed the issue with assignment.  I have it so that the variable
is searched starting from the
outermost scope to the current scope.

Fixed this for both printf(x) and printf x, and  also x = 1, for instance.