I made some good progress.
So far:

I'm implemented the LHS of the array part of the compiler.

The RHS is a bit more complex, and I ran into a bug.  I will
try to solve this bug by this weekend.

NOTE TO SELF: There is no range checking!  Try to do this
at some point!